Feature/BDN-958: Self review
