--- bidon/src/main/java/org/bidon/sdk/databinders/user/impl/appset/AppSetRepository.kt
@@ -1,8 +1,0 @@
-package org.bidon.sdk.databinders.user.impl.appset
-
-import android.content.Context
-
-internal interface AppSetRepository {
-    suspend fun getAppSetId(context: Context): String?
-    suspend fun getAppSetIdScope(context: Context): String?
-}
